\name{coriell}
\alias{coriell}
\title{Array CGH data set of Coriell cell lines}
\description{
  These are two data array CGH studies sets of Corriel cell lines taken
  from the reference below.
}
\usage{data(coriell)}
\format{A data frame containing five variables: first is clone name,
  second is clone chromosome, third is clone position, fourth (sample
  GM05296 from Web Table G) and fifth (sample GM13330 from Web Table H)
  are log2ratio for two cell lines.}
\source{
  Supplementary Information from https://www.nature.com/articles/ng754
}
\references{Snijders et al., Assembly of microarrays for genome-wide
  measurement of DNA copy number, \emph{Nature Genetics}, 2001}
\keyword{datasets}
