\name{cytoBand}
\docType{data}
\alias{cytoBand}
\title{Cytogenic band data}
\description{
  Cytogenic band data from the godenPath repository
}
\usage{data(cytoBand)}
\format{A data frame containing five variables: chromosome, start and
  end positions, band name and giesma stain.
}
\source{
  http://genome.cse.ucsc.edu/goldenPath/hg16/database/cytoBand.txt.gz
}
\keyword{datasets}
