\name{cytoBand}
\alias{cytoBand}
\title{Cytogenic band data}
\description{
  Cytogenic band data from the goldenPath repository
}
\usage{data(cytoBand)}
\format{A data frame containing five variables: chromosome, start and
  end positions, band name and giesma stain.
}
\source{
  http://hgdownload.cse.ucsc.edu/goldenPath/hg19/database/cytoBand.txt.gz
}
\keyword{datasets}
